test/testCar/drive/requirements.txt
numpy<2
pygame-ce>=2.5.0,<2.6.0
crccheck==1.3.0
crcmod==1.7
pyserial>=3.5,<4
opencv-python-headless==4.7.0.72
